Implement adopting an existing page table for X64 (#7)
Implement that ability to create a page table structure from an existing
page table root. This is useful to be able to parse a table that is
coming from another source. This will be used in the debugger to inspect
the active page table before touching memory for memory accesses,
breakpoints, etc. This is an unsafe operation as the paging create
cannot ensure the validity of an incoming page table.

Additionally, this adds the ability to consume a x64 page table
structure and return the page table root. This is useful for testing,
but could be used in scenarios where callers may have special
requirements around installing or handling the active page table.