Add borrow_allocator API to PageTable Trait
The contract of the PageAllocator requires that it be owned by
the paging crate so that it can allocate from it, but it must be
serviced by the calling code. As such, there are usecases for the
caller to need mutable access to the PageAllocator (say to refill
the pages it owns).

This commit adds an API to allow the caller to mutably borrow the
PageAllocator.