Add is_root Parameter to PageAllocator.allocate_page
Some architectures and systems have special requirements for the
base page (e.g. x86 architectures require the base table page to
be allocated below 4GB so that when APs are started and cr3 is
written using 32 bit register moves in protected mode, it gets the
correct page table root).

This commit updates the PageAllocator.allocate_page API to indicate
when the root page is being allocated to allow a system to perform
any necessary actions on it.