Add Support for x64 Large Paging (#16)
Add support for x64 large paging. With this change, the x64 Paging
implementation will create 2MB & 1GB pages for ranges that are suitable
where existing PD/PT level tables do not already exist. When a large
page is partially remapped or unmapped, the large page will
automatically be broken up to the next level of page tables. e.g. 1GB
will become 2MB and so on as needed.

This change only supports initial creation and splitting of large pages,
it does not support merging existing mappings into large pages. That is,
if a caller maps half of what could be a large page, then maps the rest
later, this will not result in a large page creation.

In support of large pages, the following changes were made:
- Unified the x64 PTE definition into the commonly used bits across
levels
- Removed the specification of 4KB in paging types.
- Added atomic swap of PTEs during edits (to avoid torn state)