Add Self-Map for AARCH64 and X64
This commit adds the self-map and zero VA range to AARCH64 and
X64 and removes the need for page table pages to be preallocated
for this crate. With the self map, we can access the page table
pages via the self map instead of identity mapping.

This does restrict the identity mappable address space on AARCH64,
due to its TTBR0/1 system. This will be resolved in a future
version of this crate. For now, the crate panics if it hits an
address that would conflict with the zero VA or self map.