x64 Paging: Move Self Map Init to Page Table Creation
Currently, the self map and zero VA are set up in install_page_table()
but there is no reason not to do it when we instantiate a new page
table and in fact it lets us clean up the code a bit.

Move the self map init and update the unit tests.