Remove the external API for predicting page count (#50)
## Description

With the self-map it should no longer be necessary for callers to
predict the number of pages needed and doing so leads to over-allocation
as the prediction logic is not considering existing mappings. This
change removes the external use of the functionality, but retains the
internal code for tests.

- [x] Impacts functionality?
- [ ] Impacts security?
- [x] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Unit tests

## Integration Instructions

Consumers must remove use of `get_page_table_pages_for_size`.