Return Error When Attempting to Map Already Mapped Page
The current code asserts when map_memory_region_internal
attempts to map a page that is already mapped. This leads
to an unuseful state for consumers, who don't know what
code led to this failure.

This patch updates the code to return an error so that the
caller can handle it and log/assert at that level.

It also introduces a new error code to return from this crate
when a range passed in is inconsistently mapped, i.e. some
entries are mapped and some are unmapped.