Update query_memory_region to Return New Error Code
Today query_memory_region returns NoMapping when any entry
in the range is unmapped. However, this is not sufficient for
callers, who may use query_memory_region to decide whether to
map a region or not, which is only valid if the entire region
is unmapped.

This patch leverages the new InconsistentMappingAcrossRange error
code to indicate that the given range has some entries that are
mapped and some that are unmapped.

It updates the query_memory_region_internal API to achieve this
goal by introducing a new enum that allows the recursive code to
identify if the current mapping state is in line with previous
mappings across other blocks.