Add repo contribution and compliance files (#60)
## Description

Closes [#74
(uefi-dxe-core)](https://github.com/OpenDevicePartnership/uefi-dxe-core/issues/74)

Adds the following files to define code of conduct, the contribution
process, and security vulnerability process within ODP.

CONTRIBUTING.md is slightly modified for UEFI contributors.

Relicensing code under "MIT" is a todo but expected to occur, so that
license is left in the file.
- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

- `cargo make all`

## Integration Instructions

- N/A

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>