Remove use of explicit registers and reduce asm usage (#77)
Explicit use of registers in `asm!` statements can cause issues if not properly declared. This was observed when the aarch64 `zero_pages` function was inlined and caused a variable being tracked in x1 to be dropped due to that routines explicit and undeclared use of x1. This resulted in an invalid address being dereferenced inconsistently.

Additionally, this change reduces the overall use of explicit `asm!` usage for readability and safety.