Revert Change to validate_address_range
During a cleanup PR, logic was moved from map/remap/unmap to
validate_address_range to ensure that the address passed in
was not greater than the max VA. However, it was not valid
to do this because query also shares that function and it
is valid to query beyond the max VA: the debugger must do
this when trying to read page tables in the self map. This
moves this back to the original locations and unblocks the debugger.