Remove Explicit Panics
As patina-paging is maturing, it should no longer be
panicking except in truly inoperable cases. This commit
removes all explicit panics (and some implicit ones doing
VA calculations) and does proper error handling.

The only explicit panics left are if a 4 level paging system
encounters attempts to operate on a 5 level PTE or for
aarch64 if the current EL is not EL1 or EL2.