x64: Prevent modified `rdi` in PageTableArchX64.zero_page()
Per Rust inline assembly docs (https://doc.rust-lang.org/reference/inline-assembly.html#operand-type),
`in` states:

  - The allocated register will contain the value of <expr> at the
    start of the assembly code.
  - The allocated register must contain the same value at the end of
    the assembly code (except if a lateout is allocated to the same register).

This code:

```rust
    unsafe fn zero_page(base: VirtualAddress) {
        let _page: u64 = base.into();
        #[cfg(all(not(test), target_arch = "x86_64"))]
        unsafe {
            asm!(
            "cld",              // Clear the direction flag so that we increment rdi with each store
            "rep stosq",        // Repeat the store of qword in rax to [rdi] rcx times
            in("rcx") 0x200,    // we write 512 qwords (4096 bytes)
            in("rdi") _page,    // start at the page
            in("rax") 0,        // store 0
            options(nostack, preserves_flags)
            );
        }
    }
```

Will capture the `_page` address at the start of the assembly code
into `rdi`. However, it uses `rep stosq` which is going to:

- Store the `rax` value in `[rdi]` (`0`)
- Repat the quad word copy `rcx` times (`0x200`)
- For each operation, `rdx` will be incremented by `8` (qword)
  - Going forward due to `cld`
- Decrement `rcx` until it is zero

At the end, `rdx` is now `_page` + (`rcx` * 8) = `_page` + `0x1000`

I think it then makes sense, to manually restore the `rdi` value
since the code is inlined. This uses `r8` with `out` to let the
compiler know the register value is modified in the asembly block.

Note: The `inline(never)` attribute in the aarch64 implementation is
removed as well as that was mentioned as being added in response to
the x64 issue as opposed to a specific aarch64 issue.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>