Move zero_page Implementation Back to Rust Code
Commit bcc008db3416aaa5a84775cb25435828ef06ba91 moved the
zero_page functionality to an asm function in order to
prevent the compiler from optimizing away the writes. This
was not seen here, but was seen in other instances in patina
and so fear drove the decision. However, after discussion with
a Rust compiler team, the issues with the compiler dropping
writes were elucidating and this case should not fall under it.

At the same time, several issues with using inline asm were
uncovered, mostly from how the asm was constructed and unforeseen
intersection with being inlined several layers up.

This commit moves these functions back to Rust code, using techniques
to ensure that the compiler does not optimize out these writes, namely
writing to a mutable pointer. Because the input is a u64 and the
later writes come from other structures, the compiler should not be
able to infer any lifetime and decide the write is not used.

The zero_page HAL API is maintained in order to allow AARCH64 to
invalid the cache for this address if the MMU is disabled.