Enforce canonical address ranges
This change adds a check that all ranges exist in the lower canonical address
range. This means that the API will not access high or kernel address space ranges
where the top bits are 1. This is already practically enforced at the APIs only
support identity mapping and so high VA addresses are not valid input
anyways.