Scheduled MIRI test execution (#106)
## Description

This PR makes a new workflow that runs every Saturday at midnight, that
runs all tests in the repository through MIRI.

This workflow allows us to filter the tests executed via the test-filter
parameter, however we will run all tests at once for now. If tests take
too long, we may need to run sets of tests on different runners to
complete in time.

If any tests failures are detected during a scheduled run, it will
create (or edit) an issue with the failed tests.

Here is an example of the issue created / edited:

https://github.com/Javagedes/patina-paging/issues/2

Here is an example of a test being detected and creating / editing the
issue


https://github.com/Javagedes/patina-paging/actions/runs/17107112335/job/48518795295

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Multiple different run scenarios on a fork.

## Integration Instructions

N/A