paging: Fix asm! Macro Usage
Patina usaged of the asm! macro was assuming that the
compiler would not reorder adjacent asm! blocks. However,
per the r-asm rules, the compiler does not make this guarantee:
https://doc.rust-lang.org/reference/inline-assembly.html#r-asm.rules.not-successive.

This also aligns usage of isb sy as the specified instruction synchronization
barrier instead of just isb. It is the same instruction now, but clarifying
this is helpful.

As such, all adjacent asm! blocks are merged in this commit when
required.