Represent PTs as Slices
This commit refactors patina-paging to represent page tables
as slices instead of doing pointer math to find entries. Pointer
math is now restricted to getting each table, instead of each
entry.

As part of this refactor, pagetablestore.rs and the abstraction
layer there has been removed: it is no longer needed since we
can use the slice iteration methods to parse the tables.