Workaround Rust Slice Iter Limitation
There is a limitation in Rust's slice::iter_mut() function
that causes a crash when querying the self map:
https://github.com/rust-lang/rust/issues/146911.

This commit works around that limitation by iterating on the
indices of the slice, not the iterator.

This is only needed for the query operation, because it
is only valid to query the self map, not map, remap, or
unmap it (it will explicitly return an error). However,
this is applied to all operations for consistency.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>