Rename set_present and get_present arch functions
On x64, Descriptor.set_present() and Descriptor.get_present()
are autogenerated functions from the bitflags crate for the
X64 PTE descriptor. These names are confusingly the same as what
the paging arch abstraction uses to get/set those fields. It has
not caused issues thus far, but is confusing to trace through and
leaves the possibility for weird symbol collision, so rename
the arch fields.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>