Invalidate Self Map VA of Next Level of Large Page
During the paging crate refactor in
4e57bb82e855d2a12102ad6b573c295e7918418f, we changed
the invalidating of the self map VA of the next level
of a large page we are splitting to be at the same level,
which is incorrect.

This restores the original logic to invalidate the TLB entry
for the self map VA of the next level of this large page,
which can get pulled in by speculative execution with the
incorrect mapping.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>