Calculate Next VA in Query
The self map VA is calculated from the start_va of a range
and the level it is in. However, when querying a range if
an unmapped page is encountered and there are further pages
to query, query can either crash or give incorrect results.

This is because query does a continue after that, moving to
the next PTE, however it does not move the start_va forward,
as is done if a mapped page is encountered. As such, the self
map VA is calculated based on the wrong VA. This can lead to
a crash if the self map VA of that range is not in use or to
potentially incorrect results if it is.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>