Create Permissive API
Today, patina-paging has a strict API set: if map is called
on a region that already contains some mapped pages, it will fail.
If remap is called on a region that is partially mapped, it will
fail.

This API design pushes opinion into this crate where it does not
need to be. The consumer of the crate should decide whether it
is appropriate to map a region which is partially unmapped. From
this crate's perspective, if it is told to map, it should map
the region. This helps limit failures in the crate and make it
a simple utility. The consumer is expected to enforce any policy
decisions.

As such, this changes the API set of this crate to drop remap.
Callers will call map and that will cause a region to be mapped
with the specified attributes, regardless of the current state.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>