Flush self-map TLB entries when unmapping large pages.
When a large page is mapped, there technically exists a self-map mapping
to the physical address of the large page. This should not be used, but
it can be pulled in by speculative execution. Therefore, when unmapping
a large page, we need to flush the self-map TLB entries for that region
similar to how it is flushed for splits.

Additionally, this commit splits the logic out for re-use and expliticly
checks for >2mb pages, since those could have multiple layers of self-map VAs.
In such cases, the entire TLB will be flushed.