AArch64: Enforce Device Memory Must be Non-Executable
The ARM ARM v8 section B2.7.2 states that setting any
device memory as executable is a programming error.
This has been confirmed by seeing crashes on physical
platforms from seeing speculative execution attempt
instruction fetches from device memory that was not
ready to be read.

This enforces that device memory be marked NX.

In addition, this changes the meaning of
PtError::IncompatibleMemoryAttributes because it wasn't
a good name for what it was trying to describe and
instead is better for this case. A new error type is
added for the query failure.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>